api/requirements.txt
django==2.2.1
djangorestframework==3.9.4
psycopg2==2.8.3
requests==2.22.0
